# backend/docs/swagger.yaml
definitions:
...
  database.Booking:
    properties:
      booking_date:
        type: string
      booking_id:
        type: integer
      citizen_id:
        type: string
      class:
        type: string
      flight_id:
        type: integer
      price:
        type: number
      seat_number:
        type: string
      total_passengers:
        type: integer
    type: object
...
  database.Passenger:
    properties:
      booking_id:
        type: integer
      citizen_id:
        $ref: '#/definitions/sql.NullString'
      dob:
        type: string
      email:
        $ref: '#/definitions/sql.NullString'
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      passenger_id:
        type: integer
      phone_number:
        $ref: '#/definitions/sql.NullString'
      seat_number:
        type: string
    type: object
...
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/bookings:
    get:
      consumes:
      - application/json
      description: Get all bookings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Booking'
            type: array
      summary: Get all bookings
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Create a booking
      parameters:
      - description: Booking object
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/database.Booking'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Booking'
      summary: Create a booking
      tags:
      - bookings
  /api/v1/bookings/{booking_id}:
    delete:
      consumes:
      - application/json
      description: Remove a booking
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a booking
      tags:
      - bookings
    put:
      consumes:
      - application/json
      description: Update a booking
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: string
      - description: Booking object
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/database.Booking'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Booking'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a booking
      tags:
      - bookings
  /api/v1/bookings/search/booking/{booking_id}:
    get:
      consumes:
      - application/json
      description: Get booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Booking'
      summary: Get booking by ID
      tags:
      - bookings
  /api/v1/bookings/search/citizen/{citizen_id}:
    get:
      consumes:
      - application/json
      description: Get booking by citizen ID
      parameters:
      - description: Citizen ID
        in: path
        name: citizen_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Booking'
            type: array
      summary: Get booking by citizen ID
      tags:
      - bookings
  /api/v1/bookings/search/flight/{flight_id}:
    get:
      consumes:
      - application/json
      description: Get booking by flight ID
      parameters:
      - description: Flight ID
        in: path
        name: flight_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Booking'
            type: array
      summary: Get booking by flight ID
      tags:
      - bookings
...
  /api/v1/passengers/booking/{booking_id}:
    get:
      consumes:
      - application/json
      description: Get all passengers by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Passenger'
            type: array
      summary: Get all passengers by booking ID
      tags:
      - passengers
...
swagger: "2.0"